# docs/swagger.yaml
basePath: /api/v1
host: localhost:3000
info:
  contact: {}
  description: A blog API to learn Go.
  title: Go API
  version: "1.0"
paths:
  /posts:
    get:
      consumes:
      - application/json
      description: Get details of all posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new post
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Create a post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Post deleted'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: error description'
          schema:
            additionalProperties: true
            type: object
      summary: Delete a post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get details of a particular post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get a post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'error: Invalid form submitted'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: error description'
          schema:
            additionalProperties: true
            type: object
      summary: Update a post
      tags:
      - posts
  /posts/{id}/edit:
    get:
      consumes:
      - application/json
      description: Get post details for editing by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: OK
        "500":
          description: 'error: error description'
          schema:
            additionalProperties: true
            type: object
      summary: Edit a post
      tags:
      - posts
swagger: "2.0"